# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saatvik Jain Aahar Gruh - Tiffin Service API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  phone?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  role: string;
  status: number;
  created_at?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Category {
  id: number;
  name: string;
  status: number;
}

export interface CategoryBody {
  name: string;
  status?: number;
}

export interface ProductWithCategory {
  id: number;
  name: string;
  description?: string | null;
  price: string;
  image_url?: string | null;
  category_id?: number | null;
  menu_date: string;
  category_name?: string | null;
  is_special: boolean;
  stock: number;
  status: number;
  created_at?: string;
}

export interface ProductBody {
  name: string;
  description?: string | null;
  price: string;
  image_url?: string | null;
  category_id?: number | null;
  menu_date: string;
  is_special?: boolean;
  stock: number;
  status?: number;
}

export interface SnackCategory {
  id: number;
  name: string;
  sortOrder: number;
  status: number;
  createdAt?: string;
}

export interface Inventory {
  snackId: number;
  quantity: number;
  updatedAt?: string;
}

export interface Snack {
  id: number;
  name: string;
  description?: string;
  price: string;
  imageUrl?: string;
  snackCategoryId?: number;
  weight?: string;
  status?: number;
  createdAt?: string;
  SnackCategory?: SnackCategory;
  Inventory?: Inventory;
}

export interface CartItem {
  id: number;
  cart_id: number;
  product_id?: number;
  snack_id?: number;
  quantity: number;
  status: number;
  product?: ProductWithCategory;
  snack?: Snack;
}

export interface CartWithItems {
  id: number;
  user_id: number;
  items: CartItem[];
  subtotal: string;
}

export interface AddToCartBody {
  product_id?: number;
  snack_id?: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface OrderItem {
  id: number;
  order_id: number;
  product_id: number;
  quantity: number;
  price: string;
  product_name?: string | null;
}

export type OrderWithItemsFulfillmentType =
  (typeof OrderWithItemsFulfillmentType)[keyof typeof OrderWithItemsFulfillmentType];

export const OrderWithItemsFulfillmentType = {
  DELIVERY: "DELIVERY",
  TAKE_AWAY: "TAKE_AWAY",
  DINE_IN: "DINE_IN",
} as const;

export interface OrderWithItems {
  id: number;
  user_id: number;
  customer_name?: string | null;
  customer_phone?: string | null;
  delivery_area_id?: number | null;
  delivery_area_name?: string | null;
  delivery_address?: string | null;
  delivery_charge: string;
  subtotal: string;
  total: string;
  status: string;
  payment_method?: string | null;
  fulfillment_type: OrderWithItemsFulfillmentType;
  order_time: string;
  delivery_date?: string | null;
  items: OrderItem[];
}

export type OrderTrackFulfillmentType =
  (typeof OrderTrackFulfillmentType)[keyof typeof OrderTrackFulfillmentType];

export const OrderTrackFulfillmentType = {
  DELIVERY: "DELIVERY",
  TAKE_AWAY: "TAKE_AWAY",
  DINE_IN: "DINE_IN",
} as const;

export interface OrderTrack {
  id: number;
  status: string;
  order_time: string;
  delivery_date?: string | null;
  delivery_area_name?: string | null;
  fulfillment_type: OrderTrackFulfillmentType;
  delivery_address?: string | null;
  items: OrderItem[];
  total: string;
}

export type CreateOrderBodyFulfillmentType =
  (typeof CreateOrderBodyFulfillmentType)[keyof typeof CreateOrderBodyFulfillmentType];

export const CreateOrderBodyFulfillmentType = {
  DELIVERY: "DELIVERY",
  TAKE_AWAY: "TAKE_AWAY",
  DINE_IN: "DINE_IN",
} as const;

export interface CreateOrderBody {
  fulfillment_type: CreateOrderBodyFulfillmentType;
  delivery_area_id?: number;
  payment_method: string;
  /** Required when fulfillment_type is DELIVERY (full street address). */
  delivery_address?: string;
}

export interface UpdateOrderStatusBody {
  status: string;
}

export interface DeliveryArea {
  id: number;
  name: string;
  delivery_charge: string;
  status: number;
}

export interface DeliveryAreaBody {
  name: string;
  delivery_charge: string;
  status?: number;
}

export interface Settings {
  id: number;
  order_cutoff_time?: string | null;
  maintenance_mode: boolean;
  announcement?: string | null;
  business_name?: string | null;
  contact_number?: string | null;
  status?: number;
}

export interface SettingsBody {
  order_cutoff_time?: string | null;
  maintenance_mode?: boolean;
  announcement?: string | null;
  business_name?: string | null;
  contact_number?: string | null;
}

export interface ContactBody {
  name: string;
  phone: string;
  email?: string | null;
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  phone: string;
  email?: string | null;
  message: string;
  status: number;
  created_at?: string;
}

export interface GalleryItem {
  id: number;
  name: string;
  description?: string | null;
  imageUrl: string;
  sortOrder: number;
  status: number;
  createdAt?: string;
}

export interface CreateGalleryItemBody {
  name: string;
  description?: string;
  imageUrl: string;
  sortOrder?: number;
  status?: number;
}

export interface UpdateGalleryItemBody {
  name: string;
  description?: string;
  imageUrl: string;
  sortOrder?: number;
  status?: number;
}

export interface DashboardStats {
  today_orders: number;
  pending_orders: number;
  delivered_today: number;
  revenue_today: string;
  active_customers: number;
  menu_items: number;
  recent_orders: OrderWithItems[];
}

export interface WeeklyRevenuePoint {
  day: string;
  revenue: string;
  orders: number;
}

export interface CategoryOrderCount {
  category_name: string;
  count: number;
  percentage: number;
}

export interface CreateSnackCategoryBody {
  name: string;
  sortOrder?: number;
  status?: number;
}

export interface UpdateSnackCategoryBody {
  name: string;
  sortOrder?: number;
  status?: number;
}

export interface SnackBody {
  name: string;
  description?: string;
  price: number;
  imageUrl?: string;
  snackCategoryId: number;
  weight?: string;
  status?: number;
}

export interface InventoryTransaction {
  id: number;
  snackId: number;
  type: string;
  quantity: number;
  note?: string;
  createdAt?: string;
  Snack?: Snack;
}

export type InventoryTransactionBodyType =
  (typeof InventoryTransactionBodyType)[keyof typeof InventoryTransactionBodyType];

export const InventoryTransactionBodyType = {
  ADD: "ADD",
  SALE: "SALE",
  ADJUSTMENT: "ADJUSTMENT",
} as const;

export interface InventoryTransactionBody {
  snackId: number;
  type: InventoryTransactionBodyType;
  quantity: number;
  note?: string;
}

export type ListProductsParams = {
  category_id?: number;
  /**
   * Menu day in YYYY-MM-DD (defaults to today)
   */
  menu_date?: string;
  search?: string;
};

export type AdminListProductsParams = {
  /**
   * Filter admin list by menu day in YYYY-MM-DD (defaults to today)
   */
  menu_date?: string;
};

export type AdminListOrdersParams = {
  status?: string;
};

export type ListSnacksParams = {
  snack_category_id?: number;
  search?: string;
};

export type AdminListInventoryTransactionsParams = {
  snack_id?: number;
};