# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Saatvik Jain Aahar Gruh - Tiffin Service API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "400":
          description: Order blocked (maintenance/cutoff)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/my:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List current user's orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
  /orders/track/{orderId}:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track order by ID (public)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTrack"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin - list all orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Admin - update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"

  # DELIVERY AREAS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        product_id:
          type: integer
        quantity:
          type: integer
        price:
          type: string
        product_name:
          type: string
          nullable: true
      required:
        - id
        - order_id
        - product_id
        - quantity
        - price
    OrderWithItems:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        customer_name:
          type: string
          nullable: true
        customer_phone:
          type: string
          nullable: true
        delivery_area_id:
          type: integer
          nullable: true
        delivery_area_name:
          type: string
          nullable: true
        delivery_address:
          type: string
          nullable: true
        delivery_charge:
          type: string
        subtotal:
          type: string
        total:
          type: string
        status:
          type: string
        payment_method:
          type: string
          nullable: true
        fulfillment_type:
          type: string
          enum: [DELIVERY, TAKE_AWAY, DINE_IN]
        order_time:
          type: string
          format: date-time
        delivery_date:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - user_id
        - delivery_charge
        - subtotal
        - total
        - status
        - fulfillment_type
        - order_time
        - items
    OrderTrack:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        order_time:
          type: string
          format: date-time
        delivery_date:
          type: string
          nullable: true
        delivery_area_name:
          type: string
          nullable: true
        fulfillment_type:
          type: string
          enum: [DELIVERY, TAKE_AWAY, DINE_IN]
        delivery_address:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: string
      required:
        - id
        - status
        - fulfillment_type
        - order_time
        - items
        - total
    CreateOrderBody:
      type: object
      properties:
        fulfillment_type:
          type: string
          enum: [DELIVERY, TAKE_AWAY, DINE_IN]
        delivery_area_id:
          type: integer
        payment_method:
          type: string
        delivery_address:
          type: string
          description: Required when fulfillment_type is DELIVERY (full street address).
      required:
        - fulfillment_type
        - payment_method
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    # DELIVERY AREAS
...