# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saatvik Jain Aahar Gruh - Tiffin Service API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyFulfillmentType } from "./createOrderBodyFulfillmentType";

export interface CreateOrderBody {
  fulfillment_type: CreateOrderBodyFulfillmentType;
  delivery_area_id?: number;
  payment_method: string;
  /** Required when fulfillment_type is DELIVERY (full street address). */
  delivery_address?: string;
}