# lib/api-zod/src/generated/types/orderTrack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saatvik Jain Aahar Gruh - Tiffin Service API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderTrackFulfillmentType } from "./orderTrackFulfillmentType";

export interface OrderTrack {
  id: number;
  status: string;
  order_time: Date;
  delivery_date?: string | null;
  delivery_area_name?: string | null;
  fulfillment_type: OrderTrackFulfillmentType;
  delivery_address?: string | null;
  items: OrderItem[];
  total: string;
}