# lib/api-zod/src/generated/types/orderWithItems.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Saatvik Jain Aahar Gruh - Tiffin Service API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderWithItemsFulfillmentType } from "./orderWithItemsFulfillmentType";

export interface OrderWithItems {
  id: number;
  user_id: number;
  customer_name?: string | null;
  customer_phone?: string | null;
  delivery_area_id?: number | null;
  delivery_area_name?: string | null;
  delivery_address?: string | null;
  delivery_charge: string;
  subtotal: string;
  total: string;
  status: string;
  payment_method?: string | null;
  fulfillment_type: OrderWithItemsFulfillmentType;
  order_time: Date;
  delivery_date?: string | null;
  items: OrderItem[];
}